os/requirements.txt
cbor2==5.7.0
luma.core==2.5.1
luma.oled==3.14.0
pillow==11.3.0
smbus2==0.5.0